backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Google Cloud
google-cloud-aiplatform==1.51.0
google-cloud-translate==3.15.3
firebase-admin==6.5.0
googlemaps==4.10.0

# Image processing
Pillow==10.3.0

# Caching & performance
cachetools==5.3.3

# Security
python-jose[cryptography]==3.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1
